prob renamed to stat
