+++ kmath-commons/build.gradle.kts
@@ -6,7 +6,7 @@ description = "Commons math binding for kmath"
 dependencies {
     api(project(":kmath-core"))
     api(project(":kmath-coroutines"))
-    api(project(":kmath-prob"))
+    api(project(":kmath-stat"))
     api(project(":kmath-functions"))
     api("org.apache.commons:commons-math3:3.6.1")
 }
