+++ kmath-core/build.gradle.kts
@@ -41,6 +41,6 @@ readme {
     feature(
         id = "autodif",
         description = "Automatic differentiation",
-        ref = "src/commonMain/kotlin/kscience/kmath/misc/SimpleAutoDiff.kt"
+        ref = "src/commonMain/kotlin/kscience/kmath/expressions/SimpleAutoDiff.kt"
     )
 }
