+++ kmath-stat/src/commonMain/kotlin/kscience/kmath/stat/RandomGenerator.kt
@@ -1,4 +1,4 @@
-package kscience.kmath.prob
+package kscience.kmath.stat
 
 import kotlin.random.Random
